[#19] Hiearachy code-split
Clean up, moved each part into separate files => improved code readability
