+++ src/components/TermAccordion.tsx
@@ -1,24 +1,13 @@
-import React, { useEffect, useState } from "react";
-import {
-  Box,
-  CircularProgress,
-  FormControl,
-  Paper,
-  Typography,
-  withStyles,
-} from "@material-ui/core";
-
+import React, { useState } from "react";
+import { FormControl, withStyles } from "@material-ui/core";
 import MuiAccordion from "@material-ui/core/Accordion";
 import MuiAccordionSummary from "@material-ui/core/AccordionSummary";
 import MuiAccordionDetails from "@material-ui/core/AccordionDetails";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 import RouteLink from "./RouteLink";
-
-import { ReactComponent as TallConnector } from "../assets/connector_taller.svg";
-
 import { TermInfo } from "./Hierarchy";
-import { TermBase, useTerm } from "../api/TermAPI";
-import { emptyTerm } from "./TermPage";
+import AccordionDescription from "./AccordionDescription";
+import HierarchyItem from "./HierarchyItem";
 
 const Accordion = withStyles({
   root: {
@@ -67,44 +56,6 @@ const ExpandIcon = withStyles((theme) => ({
   },
 }))(ExpandMoreIcon);
 
-const CurrentTermBox = withStyles((theme) => ({
-  root: {
-    backgroundColor: theme.palette.secondary.main,
-    marginBottom: -1,
-    minHeight: 56,
-    display: "flex",
-    paddingLeft: theme.spacing(2),
-    paddingRight: theme.spacing(7),
-    alignItems: "center",
-    paddingTop: 12,
-    paddingBottom: 12,
-  },
-}))(Paper);
-
-interface HierarchyItemProps {
-  level: number;
-}
-
-const HierarchyItem: React.FC<HierarchyItemProps> = (props) => {
-  //TODO: hide the connector for now, just show the structure
-
-  return (
-    <Box display="flex" ml={props.level * 5} mt={2}>
-      <Box style={{ position: "relative", minWidth: "28px" }}>
-        <TallConnector
-          style={{
-            visibility: "hidden",
-            position: "absolute",
-            top: "-45px",
-            zIndex: -1,
-          }}
-        />
-      </Box>
-      <Box style={{ flex: 1 }}>{props.children}</Box>
-    </Box>
-  );
-};
-
 export interface TermAccordionProps {
   level: number;
   term: TermInfo;
@@ -140,50 +91,4 @@ export const TermAccordion: React.FC<TermAccordionProps> = (props) => {
       </Accordion>
     </HierarchyItem>
   );
-};
-
-export const CurrentTerm: React.FC<TermAccordionProps> = (props) => {
-  return (
-    <HierarchyItem level={props.level}>
-      <Box
-        style={{
-          border: "1px solid rgba(0, 0, 0, .125)",
-          boxShadow: "none",
-        }}
-      >
-        <CurrentTermBox square elevation={0}>
-          <Typography variant="h6" color="textSecondary">
-            {props.term.label.cs}
-          </Typography>
-        </CurrentTermBox>
-      </Box>
-    </HierarchyItem>
-  );
-};
-
-const AccordionDescription: React.FC<TermBase> = (props) => {
-  const [description, setDescription] = useState<string>();
-  const { data = [], isLoading, isSuccess } = useTerm(props ?? emptyTerm);
-
-  useEffect(() => {
-    if (isSuccess) {
-      if (data.definition?.cs) {
-        setDescription(data.definition.cs);
-      } else {
-        setDescription("Pojem nemá definici");
-      }
-    }
-  }, [data]);
-
-  if (isLoading)
-    return (
-      <Box flex={1} display="flex" alignItems="center">
-        <CircularProgress />
-        <Typography>Načítání definice</Typography>
-      </Box>
-    );
-
-  if (isSuccess) return <Typography>{description}</Typography>;
-
-  return null;
 };
